tornado: Extract functions called from django into one module.
This makes clearer the separation of concerns.
