+++ tools/test-backend
@@ -112,6 +112,7 @@ not_yet_fully_covered = {path for target in [
     # Tornado should ideally have full coverage, but we're not there.
     'zerver/tornado/autoreload.py',
     'zerver/tornado/descriptors.py',
+    'zerver/tornado/django_api.py',
     'zerver/tornado/event_queue.py',
     'zerver/tornado/exceptions.py',
     'zerver/tornado/handlers.py',
